requirements_gpu.txt
tensorflow-gpu>=1.10.0
matplotlib>=2.1.0
